+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+### Added
+- action `purge_url`: purge a URL from the CDN cache (requires Terraform 1.14+);
+
 ## 0.17.0 - 2026-08-04
 
 ### Backwards compatibility breaks